updated s3 emulated structure to accurate version
