+++ s3Emulator/pv-validation-hub-bucket/evaluation_scripts/time-shift/system_metadata.csv
@@ -1,0 +1,12 @@
+system_id,name,azimuth,tilt,elevation,latitude,longitude,tracking,climate_type,dc_capacity
+1422,RTC-VT-Baseline,180,35,109,44.4665,-73.1014,FALSE,Dfb,6
+1423,RTC-NV-Baseline,180,35,697,36.0275,-114.9215,FALSE,BWh,6
+4901,NIST_Canopy_1,270,5,137,39.1385,-77.2155,FALSE,Cfa,242.5
+4902,NIST_Ground_1,180,20,138,39.1319,-77.2141,FALSE,Cfa,270.7
+4903,NIST_Roof_1,180,10,149,39.1354,-77.2156,FALSE,Cfa,73.7
+4901,NIST_Canopy_1,90,5,137,39.1385,-77.2155,FALSE,Cfa,242.5
+1332,NREL Parking Garage,180,16.77,1770,39.7388,-105.1732,FALSE,BSk,1153.488
+1332,NREL Parking Garage,180,60,1770,39.7388,-105.1732,FALSE,BSk,1153.488
+1229,FSEC - Zero Energy House - Lakeland,240,33.3,58,28.0392,-81.95,FALSE,Cfa,4.05
+50,SERFEast,158,45,1994.7,39.742,-105.1727,FALSE,BSk,6
+4,Sanyo1,180,40,1795.3,39.7406,-105.1774,FALSE,BSk,1
