+++ s3Emulator/pv-validation-hub-bucket/submission_files/submission_user_1/submission_1/results/data_sampling_frequency_mean_absolute_error_results.csv
@@ -1,0 +1,5 @@
+data_sampling_frequency,mean_absolute_error
+1,0.0
+15,7.956147220046986
+30,0.0
+60,0.0
