+++ s3Emulator/pv-validation-hub-bucket/submission_files/submission_user_1/submission_1/results/data_sampling_frequency_run_time_results.csv
@@ -1,0 +1,5 @@
+data_sampling_frequency,run_time
+1,306.4427716732025
+15,24.995373407999676
+30,9.972615003585815
+60,6.0556371212005615
