+++ s3Emulator/pv-validation-hub-bucket/submission_files/submission_user_1/submission_1/results/issue_mean_absolute_error_results.csv
@@ -1,0 +1,4 @@
+issue,mean_absolute_error
+Full DST,11.934220830070478
+No Issue,0.0
+Wrong Timezone,0.0
