+++ workers/environment.env
@@ -1,0 +1,1 @@
+env=local
